refactor: split out API app
