+++ src/project/urls.py
@@ -20,7 +20,7 @@ from django.urls import include, path
 
 urlpatterns = [
     path("", include("webview.urls")),
-    path("api/", include("shared.urls")),
+    path("api/", include("api.urls")),
     path("admin/", admin.site.urls),
     path("accounts/", include("allauth.urls")),
     path("debug/", include("debug_toolbar.urls")),
